+++ src/lib/components/loading-spinner/service/terra-loading-spinner.service.ts
@@ -10,7 +10,9 @@ import {
 /**
  * @author mscharf
  */
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class TerraLoadingSpinnerService
 {
     public observable:Observable<boolean>;